requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
pycparser==2.22
sqlparse==0.5.0
typing_extensions==4.11.0